Enforce braces, add whitespace, don't reset ui, status
Enforce braces around all blocks. The main culprits were
single statement ifs.

Added spacing to make the code a bit easier to read. There's
still a lot of work that needs to be done, but it's a good
start.

The ui won't reset if the user aborts the test. This is to
prevent already ran tests from disappearing.

Added a status message to display if the test is stopped,
running, or done.

The IP address will now show 0.0.0.0 until the IP test is
performed. The IP will never be reset but only updated
if the user initiates another test.